fix: moved name of provider into the banner
